requirements.txt
opencv-python-headless>=4.8.0
pillow>=9.5.0
numpy>=1.24.0
requests>=2.31.0
flask>=2.3.0
easyocr>=1.7.0
rembg>=2.0.50
anthropic>=0.8.0
python-dotenv>=1.0.0
scikit-image>=0.21.0
matplotlib>=3.7.0
beautifulsoup4>=4.12.0
html5lib>=1.1
lxml>=4.9.0
google-cloud-vision>=3.4.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.95.0